Finish adding and displaying products functinoality
We moved the homepage off of high voltage because
it wasn't actually going to be a static page, it's
dependent on the products and whatever is featured.

Currently the featured product is just whatever was
added last.

The admin panel and the homepage are both singleton
resources.

We add rating and image url methods to the products. The size
isn't specified but will be at some point soon.

Added a seed file to handle creating products for development.