Add mailchimp integration for signup lists.
Mailchimp integration is added using the gibbon gem. We use a
before_save callback to add it to the mailchimp api. I don't
particularly like this approach but it seems like the quickest way
to get email validation for the form and to minimize multiple
subscribers.

It's tested through an integration test which checks for bad emails.
Since the only email validation that happens is server side, if the
validation runs then it's integrating successfully. I'd like to stub
the requests so the tests run a little faster.